summary chart refactor to componetize filters
